+++ core/src/main/kotlin/lt/saltyjuice/dragas/chatty/v3/core/exception/InitializeAlreadyCalledException.kt
@@ -1,0 +1,5 @@
+package lt.saltyjuice.dragas.chatty.v3.core.exception
+
+class InitializeAlreadyCalledException(message: String = "super.initialize() was already called") : Exception(message)
+{
+}