+++ core/src/main/kotlin/lt/saltyjuice/dragas/chatty/v3/core/route/Callback.kt
@@ -1,9 +1,0 @@
-package lt.saltyjuice.dragas.chatty.v3.core.route
-
-/**
- * Marks that particular method in [Controller] is used as a callback.
- */
-@MustBeDocumented
-@Retention()
-@Target(AnnotationTarget.FUNCTION)
-annotation class Callback