+++ core/src/test/kotlin/lt/saltyjuice/dragas/chatty/v3/core/mock/MockControllerWithBadMiddlewares.kt
@@ -2,12 +2,14 @@ package lt.saltyjuice.dragas.chatty.v3.core.mock
 
 import lt.saltyjuice.dragas.chatty.v3.core.route.*
 
-class MockControllerWithMiddlewares(client: MockClient) : Controller<MockRequest, MockResponse>(client)
+@Before(MockBeforeMiddleware::class)
+@After(MockAfterMiddleware::class)
+class MockControllerWithBadMiddlewares() : Controller<MockRequest, MockResponse>()
 {
-    @Callback
-    @TestedBy("testMethod")
-    @BeforeRequest(MockBeforeMiddleware::class)
-    @AfterResponse(MockAfterMiddleware::class)
+    @On(MockRequest::class)
+    @When("testMethod")
+    @Before(MockBeforeMiddleware::class)
+    @After(MockAfterMiddleware::class)
     fun mockGenerator(request: MockRequest): MockResponse
     {
         return MockResponse(request.fieldName, request.fieldValue)