+++ core/src/test/kotlin/lt/saltyjuice/dragas/chatty/v3/core/mock/MockControllerWithMiddleware.kt
@@ -1,0 +1,20 @@
+package lt.saltyjuice.dragas.chatty.v3.core.mock
+
+import lt.saltyjuice.dragas.chatty.v3.core.route.*
+
+class MockControllerWithMiddleware() : Controller<MockRequest, MockResponse>()
+{
+    @On(MockRequest::class)
+    @When("testMethod")
+    @Before(MockBeforeMiddleware::class)
+    @After(MockAfterMiddleware::class)
+    fun mockGenerator(request: MockRequest): MockResponse
+    {
+        return MockResponse(request.fieldName, request.fieldValue)
+    }
+
+    fun testMethod(request: MockRequest): Boolean
+    {
+        return true
+    }
+}