+++ core/src/test/kotlin/lt/saltyjuice/dragas/chatty/v3/core/unit/ControllerTest.kt
@@ -1,9 +1,8 @@
 package lt.saltyjuice.dragas.chatty.v3.core.unit
 
-import lt.saltyjuice.dragas.chatty.v3.core.mock.MockClient
 import lt.saltyjuice.dragas.chatty.v3.core.mock.MockController
 import lt.saltyjuice.dragas.chatty.v3.core.mock.MockControllerWithoutTestMethod
-import lt.saltyjuice.dragas.chatty.v3.core.route.TestedBy
+import lt.saltyjuice.dragas.chatty.v3.core.route.When
 import org.junit.Assert
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -24,21 +23,21 @@ class ControllerTest
     fun controllerHasTestCallback()
     {
         val callback = instance.methods()[0]
-        val testCallbackName = callback.getAnnotation(TestedBy::class.java).value
+        val testCallbackName = callback.getAnnotation(When::class.java).value
         Assert.assertNotNull(instance.javaClass.methods.find { it.name == testCallbackName })
     }
 
     @Test
     fun controllerDoesntHaveSuchTestCallback()
     {
-        val callback = MockControllerWithoutTestMethod(MockClient()).methods()[0]
-        val testCallbackName = callback.getAnnotation(TestedBy::class.java).value
+        val callback = MockControllerWithoutTestMethod().methods()[0]
+        val testCallbackName = callback.getAnnotation(When::class.java).value
         Assert.assertNull(instance.javaClass.methods.find { it.name == testCallbackName })
     }
 
     companion object
     {
         @JvmStatic
-        val instance = MockController(MockClient())
+        val instance = MockController()
     }
 }